simplify code, drop compression

+++ src/lib.rs
@@ -11,7 +11,7 @@ pub fn get_title(game_id: impl AsRef<str>) -> Option<&'static str> {
 }
 
 #[cfg(feature = "gamehacking")]
-pub fn get_ghid(game_id: impl AsRef<str>) -> Option<u32> {
+pub fn get_ghid(game_id: impl AsRef<str>) -> Option<usize> {
     let game_id = u32::from_str_radix(game_id.as_ref(), 36).ok()?;
     core::get_ghid(game_id)
 }
